First release.
This repository is for those who wants to join the competition of 2013 NCU
CSIE Website Design. This commit is the first release of the templetes, an
may update in the future until the competition finished. Any one who wants
join should create a branch and request a merge to submit work. More detai
will be included in the README in next commit recently and be announced on
homepage of NCU CSIE (http://www.csie.ncu.edu.tw).